Multiple updates to lookback_rnn. Convert basic_rnn to shared format. Add download_midiworld_classical.py.
