--- magenta/models/basic_rnn/basic_rnn_create_dataset_test.py
@@ -1,131 +1,0 @@
-# Copyright 2016 Google Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Tests for encoders."""
-
-import basic_rnn_create_dataset
-import tensorflow as tf
-
-from magenta.lib import melodies_lib
-
-NO_EVENT = melodies_lib.NO_EVENT
-NOTE_OFF = melodies_lib.NOTE_OFF
-
-
-def one_hot(value, length):
-  return [1.0 if value == i else 0.0 for i in range(length)]
-
-
-class CreateDatasetTest(tf.test.TestCase):
-
-  def testBasicOneHotEncoder(self):
-    steps_per_beat = 4
-    transpose_to_key = 0
-    min_note = 48
-    max_note = 84
-    num_classes = max_note - min_note + 2
-
-    melody = melodies_lib.Melody()
-    melody.from_event_list([
-        NO_EVENT, 60, 62, 64, NO_EVENT, NOTE_OFF, 65, 67,
-        NOTE_OFF, 69, 71, 72, NO_EVENT, NOTE_OFF, 74, 76,
-        77, 79, NO_EVENT, NOTE_OFF])
-    transformed_melody = [
-        NO_EVENT, 12, 14, 16, NO_EVENT, NOTE_OFF, 17, 19,
-        NOTE_OFF, 21, 23, 24, NO_EVENT, NOTE_OFF, 26, 28,
-        29, 31, NO_EVENT, NOTE_OFF]
-    expected_inputs = ([one_hot(note + 2, num_classes)
-                        for note in transformed_melody] +
-                       [one_hot(0, num_classes)] * 12)
-    expected_labels = [note + 2 for note in transformed_melody[1:]] + [0] * 13
-    melody.squash(min_note, max_note, transpose_to_key)
-    inputs, labels = basic_rnn_create_dataset.basic_one_hot_encoder(
-        melody, steps_per_beat, min_note, max_note)
-    self.assertListEqual(inputs.tolist(), expected_inputs)
-    self.assertListEqual(labels, expected_labels)
-
-  def testBasicOneHotEncoderTruncateNoteOff(self):
-    steps_per_beat = 4
-    transpose_to_key = 0
-    min_note = 48
-    max_note = 84
-    num_classes = max_note - min_note + 2
-
-    melody = melodies_lib.Melody()
-    melody.from_event_list([
-        NO_EVENT, 60, 62, 64, NO_EVENT, NOTE_OFF, 65, 67,
-        NOTE_OFF, 69, 71, 72, NO_EVENT, NOTE_OFF, 74, 76, NOTE_OFF])
-    transformed_melody = [
-        NO_EVENT, 12, 14, 16, NO_EVENT, NOTE_OFF, 17, 19,
-        NOTE_OFF, 21, 23, 24, NO_EVENT, NOTE_OFF, 26, 28]
-    expected_inputs = [one_hot(note + 2, num_classes)
-                       for note in transformed_melody]
-    expected_labels = ([note + 2 for note in transformed_melody[1:]] +
-                       [NOTE_OFF + 2])
-    melody.squash(min_note, max_note, transpose_to_key)
-    inputs, labels = basic_rnn_create_dataset.basic_one_hot_encoder(
-        melody, steps_per_beat, min_note, max_note)
-    self.assertListEqual(inputs.tolist(), expected_inputs)
-    self.assertListEqual(labels, expected_labels)
-
-  def testBasicOneHotEncoderOctaveSquash(self):
-    steps_per_beat = 4
-    transpose_to_key = 0
-    min_note = 48
-    max_note = 84
-    num_classes = max_note - min_note + 2
-
-    melody = melodies_lib.Melody()
-    melody.from_event_list([
-        NO_EVENT, 84, 86, 52, NO_EVENT, NOTE_OFF, 65, 67,
-        NOTE_OFF, 69, 71, 72, NO_EVENT, NOTE_OFF, 38, 40, NOTE_OFF])
-    transformed_melody = [
-        NO_EVENT, 24, 26, 4, NO_EVENT, NOTE_OFF, 17, 19,
-        NOTE_OFF, 21, 23, 24, NO_EVENT, NOTE_OFF, 2, 4]
-    expected_inputs = [one_hot(note + 2, num_classes)
-                       for note in transformed_melody]
-    expected_labels = ([note + 2 for note in transformed_melody[1:]] +
-                       [NOTE_OFF + 2])
-    melody.squash(min_note, max_note, transpose_to_key)
-    inputs, labels = basic_rnn_create_dataset.basic_one_hot_encoder(
-        melody, steps_per_beat, min_note, max_note)
-    self.assertListEqual(inputs.tolist(), expected_inputs)
-    self.assertListEqual(labels, expected_labels)
-
-  def testBasicOneHotEncoderTransposeKey(self):
-    steps_per_beat = 4
-    transpose_to_key = 0
-    min_note = 48
-    max_note = 84
-    num_classes = max_note - min_note + 2
-
-    melody = melodies_lib.Melody()
-    melody.from_event_list([
-        NO_EVENT, 61, 63, 65, NO_EVENT, NOTE_OFF, 66, 68,
-        NOTE_OFF, 70, 72, 73, NO_EVENT, NOTE_OFF, 75, 77, NOTE_OFF])
-    transformed_melody = [
-        NO_EVENT, 12, 14, 16, NO_EVENT, NOTE_OFF, 17, 19,
-        NOTE_OFF, 21, 23, 24, NO_EVENT, NOTE_OFF, 26, 28]
-    expected_inputs = [one_hot(note + 2, num_classes)
-                       for note in transformed_melody]
-    expected_labels = ([note + 2 for note in transformed_melody[1:]] +
-                       [NOTE_OFF + 2])
-    melody.squash(min_note, max_note, transpose_to_key)
-    inputs, labels = basic_rnn_create_dataset.basic_one_hot_encoder(
-        melody, steps_per_beat, min_note, max_note)
-    self.assertListEqual(inputs.tolist(), expected_inputs)
-    self.assertListEqual(labels, expected_labels)
-
-
-if __name__ == '__main__':
-  tf.test.main()
